Fix bug during the saving.
If an option value is only composed of whitespaces (space or tabulation), we
have to quote them during the saving process.
If we do not do that, then when we will try load the file, the parser will
remove them (because it will treat them as useless blank) and we will get an
empty value.