Fix bug during the saving.
If an option value starts like a comment (with ; or #), we have to quote it
during the saving process.
If we do not do that, then when we will try load the file, the parser will
remove them (because it will treat them as comments) and we will get an
empty value.